Convert parts into modules
