+++ package.json
@@ -17,9 +17,7 @@
     "espree": "6.1.2",
     "esprima-walk": "^0.1.0",
     "js-beautify": "^1.9.0",
-    "lodash": "^4.17.11",
-    "sha1": "^1.1.1",
-    "silent-error": "^1.1.1"
+    "lodash": "^4.17.11"
   },
   "devDependencies": {
     "@ember/optional-features": "^0.6.3",
